Get the Geometry Data for a File.

Get the geometry of an uploaded file using the Model Derivative API.

Usage

getData(guid = NULL, urn = NULL, token = NULL)

Arguments

guid

A string. GUID retrieved via the getMetadata function.

urn

A string. Source URN (objectId) for the file. Note the URN must be Base64 encoded. To encode the URN, see, for example, the jsonlite::base64_enc function.

token

A string. Token generated with getToken function with data:read and data:write scopes.

Value

An object containing the geometry data for the selected file.

Examples

if (FALSE) { # \dontrun{
# Get the geometry data for the "aerial.dwg" svf file
resp <- getData(guid <- myGuid, urn <- myEncodedUrn, token = myToken)
} # }